move purge script to correct place
